Data handling improvements
Fixes a major bug with timestamps (the data was being indexed by timestamps, so any timestamps beyond the maximum time range would not work). Now, timestamps will function like actual time and be spaced apart on the graph according to their values.

Introduces the `activity` prop for LineGraph to designate which lines are active. This is more in line with how real data will be fetched. However, it is a required prop which is annoying if a user of LineGraph doesn't need that functionality.

Merges all experiment data into one file, which should simplify the data fetching process.