Filter x-axis instead of zoom
Allows using the zoom feature on the x-axis while still filtering based on the global time range. The zoom is not very smooth because it throttles its frame rate - the throttling can be turned off, but we'll see how real data performs first.

The tooltip is broken, but using axis indicators will be a better solution (16 wells in one tooltip is a lot).